autocommit when autocommit are off.
according to django sqlite has bugs where it does not handle
transactions and save points correctly in autocommit mode and thus
autocommits are disabled. and therefore issues a begin statement
to explicitly disable autocommit while going through transaction/atomic
block.

However this seems not to be the case in go-sqlite3 as a result we can
disable this special handling for sqlite3 only.

other databases do not have this breakage (as django call it) and
therefore do not set this flag

uncommenting the BEGIN command is less as the function is never called
becasuse the other flag is set to false (default)