rely on sqlite implementation of django
Removed almost all of the code and relying completely on the sqlite3
implementation of django.

Django communicates via the DatabaseWrapper class of the datatbase
backend so just importing it should get django to believe that the
backend is loaded and ready

So only the basic functions are only defined to create a connection and
execute by replacing the paramstyle format ('format' by 'qmark')

- Empty _start_transaction_under_autocommit function
according to django sqlite has bugs where it does not handle
transactions and save points correctly in autocommit mode and thus
autocommits are disabled. and therefore issues a begin statement
to explicitly disable autocommit while going through transaction/atomic
block.

However this seems not to be the case in go-sqlite3 as a result we can
disable this special handling for sqlite3 only.